merge-ort: replace string_list_df_name_compare with faster/clearer alternative
This function surprisingly showed up as a significant cost in the perf
profiles.  It was a little difficult to understand, particularly with
the copy-pasted rationale from merge-recursive that used the function
for a different purpose.  And there were a couple levels of indirection.
Simplify the code and speed it up by about 25%.

It still feels too long, but anything else is going to be a bit harder.
Perhaps we can keep some kind of ordering information from when we run
through traverse_trees, as that would allow a smaller comparison than a
full string comparison.  But it runs slightly counter to my idea of
allowing more trivial-ish tree merging.

Signed-off-by: Elijah Newren <newren@gmail.com>