diffcore-rename: simplify and accelerate register_rename_src()
reigster_rename_src() took pains to create an array in rename_src which
was sorted by pathname of the contained diff_filepair.  However, the
fact that this array was sorted was not needed anywhere, and thus
represented wasted time.  Simply append to the end of the array, to save
45% of diffcore_rename_extended() setup time, and 2% of overall
execution time of rebasing a series of commits with lots of renames.

Technically, there is one difference in the end result for the code.  IF
the caller of diffcore-rename provides a set of pairs with a duplicate
filename in the sources (an erroneous input condition), the old code
would simply ignore the duplicate (without warning or error).  The new
code will simply swallow it and thus allow multiple pairings for the
"same" source file.  Checking for this error condition is expensive
(basically undoing the optimization) and the only callers are the
diffcore and merge-ort machinery which are already programmed to
correctly avoid passing invalid inputs.  If someone wants to add such
error checking, I believe they should add a separate function which
callers can call before calling diffcore_rename[_extended]().

Signed-off-by: Elijah Newren <newren@gmail.com>