Accelerate rename_dst setup
The code in diffcore-rename simply makes a copy of the existing pair for
rename_src, but for rename_dst, for some reason it copies the second
diff_filespec from the existing pair and then when it finds a pairing,
it creates a new diff_filepair to stick in rename_dst for it.  This
makes it somewhat inconsistent with rename_src, as well as requiring
more memory.  Let's use the original diff_filepair in rename_dst just as
we do with rename_src.

Doing this has the side-effect that we don't have to look up the pair
from rename_dst when we are going through pairs in the original queue,
because they are the same pair.  So we can just push it on the queue
when we are dealing with a destination pair without doing any lookup.
That saves us some time and simplifies the code.

In fact, the two things above mean that the only remaining reason for
locate_rename_dst() is for handling broken pairs (i.e. when break
detection is on).  Those are somewhat rare, but we can set up a simple
strmap to get the map between the source and the index.  Doing so allows
us to avoid sorting the rename_dst array, which was done in a weakly
quadratic manner.

Overall, this accelerates setup time within diffcore-rename by about
65%, and final write back to the output queue time by about 50%, with an
overall savings of about 3.5%.