fast-export: make handle_tail() more robust
In handle_tail(), we always process commits from the end of the
object_array to the beginning.  When a commit in the commits
object_array could not be handled yet, we did two things: place that
commit back at the end of the array, and bail on handling any more
commits from the array.  This had two implicit assumptions buried in it:

  * No commit in the object_array will depend upon any commit that was
    added to the array before it.
  * For as long as there are still commits in the object_array, there
    will be another call to handle_tail() later to process them.

The first assumption arises because processing from the end to beginning
would cause us to keep bailing every time we hit a commit that depended
on one in the array before it.  This first assumption was only valid
because our "is-a-dependency-criteria" has been is-a-parent, and the
ordering of commits returned by revision walking happens to return them
in an order that will work with our assumption.  If we were to add any
other criteria, though, then we'd find this assumption likely to be
violated.  (As an example, if a cherry-pick references a commit from a
different branch in its commit message, and we want to process the
cherry-pick after the cherry-picked commit, adding this dependency
criteria to has_unshown_parent() would case handle_tail() to no longer
work.)

It is also clear that if the first assumption is violated, but the
second assumption holds, that you'd end up with an infinite loop.
However, the second assumption tends to fail exactly when the first
assumption does.  Although handle_tail() was called from
cmd_fast_export()'s loop over get_revision(), that did not help with the
guarantees at all; the handle_tail() call could have been moved outside
that loop with no change in overall behavior.  In other words, the
overall structure only worked if the last time handle_tail() was called
there were no more elements that need to be added to the object_array
and all commits in the array happened to satisfy the first assumption.
Otherwise, the current code structure would just drop all remaining
commits from the object_array and never process them at all.

We want to make handle_tail() continue operating on commits for as long
as some still remain in the array, but to make that code change we have
to make sure the object array has been populated with all necessary
commits first.  To that end, move the handle_tail() call out of the
get_revision() loop, and then add an extra loop layer in handle_tail()
to iterate until all commits in the array are processed.

Two things worth pointing out with this transformation:

  - the result definitely appears to be theoretically O(N^2), but if the
    outer loop runs more than once with the current code, then the
    previous implementation was busted and would not process all the
    commits.  In other words, the current definition of dependencies and
    guarantees of get_revision() mean that this function in practice
    is currently only O(N).  If has_unshown_parent() or get_revision()
    change, depending on how much they change, an alternate
    implementation (perhaps a priority queue?) could be useful here.

  - if the dependency condition (currently has_unshown_parent()) ever
    results in a cyclic dependency, then this code would yield an
    infinite loop.  Since parents can't have cycles this shouldn't be a
    problem, but if has_unshown_parent() is changed to allow cycles then
    additional work might also be needed in handle_tail() to know what
    to do with those cycles.

Signed-off-by: Elijah Newren <newren@gmail.com>