merge-ort: also defer recursing into new directories
In the previous commit, we deferred and sometimes avoided recursing into
directories when one side matched the merge base.  We can do the same
for directories added on one side of history (which are quite similar,
because you can view a non-existent directory in the merge base and the
other side of history as both having a null_oid and thus matching).

Signed-off-by: Elijah Newren <newren@gmail.com>