attempt to measure malloc/calloc/realloc/free overhead
Unfortunately, the overhead of using getnanotime() seems to be pretty
high itself, making accurate measurement difficult.  My results were
as follows:

    malloc_time:  0.145553
    calloc_time:  0.115598
    realloc_time: 0.031338
    free_time:    0.488756
    overhead_time: 0.705435

    Minus estimated clock measurement overhead:
    malloc_time:  0.112696
    calloc_time:  0.089503
    realloc_time: 0.024264
    free_time:    0.378424
    total:        0.604887

In order to get the overhead_time, I had to make extra calls to
getnanotime(), further pushing the amount of overhead involved.
However, since the run took 3.640s, and 3.640-.705435 = 2.934565, which
is kind of close to 2.8 (roughly the time it'd take without this
measurement overhead), I think my overhead_time approximation is fairly
decent.

Using 2.8 as the estimate of what the real time is, I'd come up with
these percentages for my hwmon-updates rebase testcase in terms of
overall time usage:
    malloc_time:   4.0%
    calloc_time:   3.2%
    realloc_time:  0.9%
    free_time:    13.5%
    total:        21.6%

Signed-off-by: Elijah Newren <newren@gmail.com>