strmap, merge-ort: always strdup strmap strings
This requires some special care since merge-ort assumes it can compare
path strings by comparing their pointers.  This works in merge-ort
because all path strings come from collect_merge_info's walk of that
paths...plus paths created from directory rename detection.

The biggest issue here is that when we add a path to &opt->priv->paths
(or any other strmap), we get a copy of that path.  Thus, we have to be
careful to use the value from the &opt->priv->paths strmap, not one we
had before or one we got from a different strmap/strset/strintmap.

Signed-off-by: Elijah Newren <newren@gmail.com>