merge tests: expect improved directory/file conflict handling in ort
merge-recursive.c is built on the idea of running unpack_trees() and
then "doing minor touch-ups" to get the result.  Unfortunately,
unpack_trees() was run in an update-as-it-goes mode, leading to
merge-recursive.c to follow suit and end up with an immediate evaluation
and fix-it-up-as-you-go design.  Some things like directory/file
conflicts are not well representable in the index data structure, and
required special extra code to handle.  But then when it was discovered
that rename/delete conflicts could also be involved in directory/file
conflicts, the special directory/file conflict handling code had to be
copied to the rename/delete codepath.  And for modify/delete, and for
rename/rename(1to2) conflicts.  And then it still missed some.  Further,
when it was discovered that there were also file/submodule conflicts and
submodule/directory conflicts, we needed to copy that code to all the
special cases.

And then it was discovered that our handling of directory/file conflicts
was suboptimal because it would create untracked files to store the
contents of the conflicting file, which would not be cleaned up if
someone were to run a 'git merge --abort' or 'git rebase --abort'.  It
was also difficult or scary to try to add or remove the index entries
corresponding to these files given the directory/file conflict in the
index.  But changing it was a royal pain because there were so many
callsites and so many tests littered over the testsuite that would need
updating.

merge-ort works hard to push all directory/file/submodule conflict
handling through a single codepath, and avoids creating untracked files
for storing tracked content (it does record things at alternate paths,
but makes sure they have higher-order stages in the index).

Since updating merge-recursive is too much work and we don't want to
destabilize it, instead update the testsuite to have different
expectations for relevant directory/file/submodule conflict tests.

Signed-off-by: Elijah Newren <newren@gmail.com>