merge tests: expect slight differences in output for recursive vs. ort
The ort merge strategy has some slight differences in commit descriptions
(shortened hashes), stdout vs stderr, and in conflict messages.  Also,
builtin/merge.c reports usage of "ort" as "Merge made by the 'ort'
strategy" -- while it is meant as a drop in replacement for "recursive"
it is not yet treated as though it is recursive.  Update the testcases
to expect different output for the different merge backends.

Signed-off-by: Elijah Newren <newren@gmail.com>