t6423: note improved ort handling with untracked files
The "recursive" backend relies on unpack_trees() to check if unstaged or
untracked files would be overwritten by a merge, but unpack_trees() does
not understand renames -- and once it returns, it has already written
many updates to the working tree and index.  As such, "recursive" had to
do a special 4-way merge where it would need to also treat the working
copy as an extra source of differences that we had to carefully avoid
overwriting and resulting in moving files to new locations to avoid
conflicts.

The "ort" backend, by contrast, does the complete merge inmemory, and
only updates the index and working copy as a post-processing step.  If
there are untracked files in the way, it can simply abort the merge.

Update t6423 to reflect the better merge abilities and expectations for
ort, while still leaving the best-case-as-good-as-recursive-can-do
expectations there for the recursive backend so we retain its stability
until we are ready to deprecate and remove it.

Signed-off-by: Elijah Newren <newren@gmail.com>