merge-ort: handle book-keeping around two- and three-way content merge
In addition to the content merge (which will go in a subsequent commit),
we need to worry about conflict messages, placing results in higher
order stages in case of a df_conflict, and making sure the results are
placed in ci->merged.result so that they will show up in the working
tree.  Take care of all that book-keeping and introduce a new
handle_content_merge() function with a overly simplistic just-take-HEAD
implementation for now.

Signed-off-by: Elijah Newren <newren@gmail.com>