merge-ort: record the reason that we want a rename for a diretory
When a directory is removed on one side of history, and a new file is added
to that directory on the other side of history, we need to detect a
potential rename of the directory so that the new file can be moved to
the right directory.  However, diffcore-rename can take advantage of the
difference between whether the current directory had a new file added or
only some ancestor of the current directory had a new file added.  (And,
obviously, can take advantage of whether neither the current directory
nor any ancestor had a new file added to it.)  Save this extra
information in dirs_removed.

Signed-off-by: Elijah Newren <newren@gmail.com>