merge-ort: reuse path strings in pool_alloc_filespec
pool_alloc_filespec() was written so that the code when pool != NULL
mimicked the code from alloc_filespec(), which including allocating
enough extra space for the path and then copying it.  However, the path
passed to pool_alloc_filespec() is always going to already be in the
same memory pool, so we may as well reuse it instead of copying it.

For the testcases mentioned in commit 557ac0350d ("merge-ort: begin
performance work; instrument with trace2_region_* calls", 2020-10-28),
this change improves the performance as follows:

                            Before                  After
    no-renames:       200.8 ms ±   3.8 ms    200.7 ms ±   4.5 ms
    mega-renames:     893.2 ms ±  21.8 ms    853.0 ms ±   7.5 ms
    just-one-mega:    271.7 ms ±   2.1 ms    263.7 ms ±   3.1 ms

Signed-off-by: Elijah Newren <newren@gmail.com>