merge-ort: mark a few more conflict messages as omittable
path_msg() has the ability to mark messages as omittable when recording
conflict messages in an in-tree file.  This conflict message file will
then be shown by diffing the merge-created tree to the actual merge
commit that is created.  Experience with the feature suggests that while
these messages are very useful when trying to create a merge initially,
they are just noise when trying to see what changes users made to create
the merge commit.

Signed-off-by: Elijah Newren <newren@gmail.com>