show, log: provide --remerge-diff-only capability
--remerge-diff-only is similar to --remerge-diff and implies its
behavior, but additionally will also change how all single-parent
commits are displayed.

If a single parent commit cannot be determined to be either a
cherry-pick or a revert (based on its commit message), then: (a) for
`git log`, the commit will simply be skipped, (b) for `git show`, a
warning that we could not find cherry-pick or revert information for the
commit will be shown.

If a single parent commit can be determined to be a cherry pick or a
revert (and the commit it is a cherry pick or revert of can be found in
the commit message), then --remerge-diff-only will cause that commit to
be automatically reverted or cherry-picked again (but only creating a
new tree, not a new commit), and then the commit will be diffed against
the automatic revert or cherry-pick tree.

Signed-off-by: Elijah Newren <newren@gmail.com>