sequencer: remove unnecessary variable setting
revs.pretty_given is only ever used in builtin/log.c, which is never
called from sequencer.  So setting this variable cannot do anything.
Remove it.

This was introduced in commit 62db524779 ("rebase -i: generate the
script via rebase--helper", 2017-07-14), which was converting from a
literal usage of log, which suggests this line was copied or ported
from log.c even though it has no use here.