spanhash_cmp: implement as external c lib for performance testing
In linux.git, running:
  ../git/misc/summarize-perf /usr/bin/git diff -l0 --dirstat -C v2.6.22..v2.6.23

With Vanilla git (NO_RUST=1):
    4.919109 : 1 : label:inexact renames
With spanhash_cmp reimplemented in Rust (no extra variables):
    4.969647 : 1 : label:inexact renames
With spanhash_cmp put into a separate .a but still in C
(KINDA_LIKE_RUST=1):
    4.965963 : 1 : label:inexact renames

Where the timings above in the left column (in seconds) represent the
best timing of three runs.  The differences are pretty close to the
run-to-run variability and the latter two are only about 1% slower, but
over many runs I was seeing the Rust code on average be slower than the
C code, leading me to wonder if I had translated it correctly.  I wanted
to make sure I wasn't doing something dumb in Rust and that there
weren't extra checks I needed to turn off somehow to get similar code.
The fact that merely moving a C function to a different .a file yields
virtually identical results means the slowdown can likely be attributed
just to the code locality.  In any event, I'm matching the performance
of the equivalent C, so I'm happy now.

Signed-off-by: Elijah Newren <newren@gmail.com>