xdiff/histogram: narrow fix for suboptimal delete-and-replace diff
A few years back, Stefan reported an issue with histogram diffs at [1].
I will modify the testcase just slightly in this description for
simplicity of explanation.  The problem arises in the step where
histogram diff looks for the longest region of matching lines
surrounding a unique-ish line.  For a testcase where the
tied-for-longest sequence was a move from the end of the file to the
beginning, e.g.

   seq 1 9 >f1
   (echo 9; seq 1 2 7) >f2
   git diff --no-index --histogram f1 f2

histogram would match up the '9' from the end of f1 to the '9' at the
start of f2.  This would mean treating everything before the '9' in f1
as deleted, and treating everything after the '9' in f2 as added.  While
a valid diff, it is rather suboptimal.  There were four other one-line
regions that were preserved, and our choice to preserve the '9' prevents
us from matching up four other lines.

One thing to note here was that we had a five-way tie for "longest"
sequence -- each being a single line long.  Picking one of the others
that tie would have given us a much nicer diff.  And in particular,
picking a choice where the two matches are nearer the beginning of each
file would give us more chances for a match.  This patch does so,
basically shifting the tiebreaker from "region with earliest start in
the post-image file" (though that tiebreaker is non-obvious and due to
order of iteration that try_lcs() is called) to "region with earliest
average start in the pre-image and post-image files".  This fixes the
particular testcase, but somewhat narrowly.  It is pretty easy to see
how to modify the testcase to still get a suboptimal result, but that
will be a discussion for the next patch...

[1]
https://lore.kernel.org/git/CAGZ79kZYO6hHiAM8Sfp3J=VX11c=0-7YDSx3_EAKt5-uvvt-Ew@mail.gmail.com/

Reported-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Elijah Newren <newren@gmail.com>