xdiff/histogram: more thorough fix for suboptimal diff
TLDR ("Too Long; Didn't Read" summary):
  * Use combined weight of sequential regions instead of just the single
    biggest one.
  * Histogram diff is fundamentally different than patience diff, not
    an extension of it.  But we make it slightly more similar.

In more detail...

In the previous commit we provided a narrow fix for a pathological case
where histogram provides a bad diff:

    seq 1 9 >f1
    (echo 9; seq 1 2 7) >f2
    git diff --no-index --histogram f1 f2

However, that only worked if the region moved to the front was tied for
the biggest single region.  If the region moved to the front was
slightly larger, then we'd again be faced with a diff that mostly
deleted lines from the first file and added lines from the second.  In
particular, a change like the following would trigger it:

    (echo 9; echo 10; seq 1 2 7) >f2

The problem here is that the other four lines could be shown as matching
if we preferred them, but we have to consider them in aggregate or else
the '9' and '10' lines is a single region that is bigger.  So, we would
like to have some kind of aggregate weighting for regions.  Would adding
such an aggregate weighting make sense to include as part of histogram
diff, or should the algorithm be renamed if we do that?  Let's dig in a
bit...

Our histogram diff was ported from JGit, where it was described as[1]:

    An extended form of Bram Cohen's patience diff algorithm.

    This implementation was derived by using the 4 rules that are
    outlined in Bram Cohen's blog, and then was further extended to
    support low-occurrence common elements.

    ...

    this algorithm behaves exactly like Bram Cohen's patience diff
    whenever there is a unique common element available between the two
    sequences.

Although the histogram diff was intended to be an extension of patience
diff, it failed to achieve that in practice.  The biggest difference is
actually pretty glaring when you dig in: patience diff was specifically
designed around a longest common subsequence problem (one that involves
the unique lines and the line numbers they are found on), and got its
name from using a "patience sort" to find the answer to the longest
subsequence of unique lines that could be matched up.  Everything in the
patience diff algorithm revolved around that.  Histogram diff talks
about "longest common subsequence", but never sets up or solves any such
problem as part of its algorithm.  It appears to have confused "longest
common substring" or "longest common sequence" for "longest common
subsequence".  (Note that this is an easy mistake to make, as evidenced
by the first sentence of the Wikipedia entry for "longest common
subsequence").

Histogram has another difference, too, in that it looks for matching
lines around unique lines, gathering things up into matching regions.
Histogram diff simply picks the biggest single such region (though the
tie-breaker changed in the last commit from the earliest one in the
post-image to the one with the earliest average location between
pre-image and post-image).  In contrast, patience used a region size of
1 (just the unique line itself) but then looked for the longest common
subsequence of such lines.  I think the extended region size that
histogram diff uses is beneficial; there are cases where histogram gives
a better diff than patience, and I think it can be attributed to this
factor.  But histogram's focus on the single largest region is a clear
drawback, as evidenced by the bad testcase above.

Using patience as inspiration, we could perhaps set up a weighted
longest common subsequence problem.  However, a much simpler (and
faster) fix is to just look for the largest uninterrupted set of
regions.  In particular, while scanning for regions, if the next region
we find comes after the previous one in both the pre-image and
post-image files, then we can just combine their weights.

<In git.git:>
[newren@iron-rod git (histogram-improvements)]$ time hyperfine -L a histogram,minimal,myers,patience 'git log --oneline --shortstat --diff-algorithm={a} -3000 v2.39.1 >/dev/null'
Benchmark 1: git log --oneline --shortstat --diff-algorithm=histogram -3000 v2.39.1 >/dev/null
  Time (mean ± σ):      2.182 s ±  0.008 s    [User: 2.022 s, System: 0.147 s]
  Range (min … max):    2.167 s …  2.195 s    10 runs

Benchmark 2: git log --oneline --shortstat --diff-algorithm=minimal -3000 v2.39.1 >/dev/null
  Time (mean ± σ):      5.385 s ±  0.028 s    [User: 5.204 s, System: 0.159 s]
  Range (min … max):    5.350 s …  5.424 s    10 runs

Benchmark 3: git log --oneline --shortstat --diff-algorithm=myers -3000 v2.39.1 >/dev/null
  Time (mean ± σ):      2.357 s ±  0.010 s    [User: 2.181 s, System: 0.164 s]
  Range (min … max):    2.336 s …  2.374 s    10 runs

Benchmark 4: git log --oneline --shortstat --diff-algorithm=patience -3000 v2.39.1 >/dev/null
  Time (mean ± σ):      2.194 s ±  0.007 s    [User: 2.040 s, System: 0.142 s]
  Range (min … max):    2.185 s …  2.205 s    10 runs

Summary
  'git log --oneline --shortstat --diff-algorithm=histogram -3000 v2.39.1 >/dev/null' ran
    1.01 ± 0.01 times faster than 'git log --oneline --shortstat --diff-algorithm=patience -3000 v2.39.1 >/dev/null'
    1.08 ± 0.01 times faster than 'git log --oneline --shortstat --diff-algorithm=myers -3000 v2.39.1 >/dev/null'
    2.47 ± 0.02 times faster than 'git log --oneline --shortstat --diff-algorithm=minimal -3000 v2.39.1 >/dev/null'

<In linux.git:>
$ time hyperfine -L a histogram,minimal,myers,patience 'git log --oneline --shortstat --diff-algorithm={a} -10000 v6.0 >/dev/null'
Benchmark 1: git log --oneline --shortstat --diff-algorithm=histogram -10000 v6.0 >/dev/null
  Time (mean ± σ):     10.368 s ±  0.167 s    [User: 9.766 s, System: 0.523 s]
  Range (min … max):   10.182 s … 10.715 s    10 runs

Benchmark 2: git log --oneline --shortstat --diff-algorithm=minimal -10000 v6.0 >/dev/null
  Time (mean ± σ):     11.106 s ±  0.116 s    [User: 10.656 s, System: 0.366 s]
  Range (min … max):   10.957 s … 11.308 s    10 runs

Benchmark 3: git log --oneline --shortstat --diff-algorithm=myers -10000 v6.0 >/dev/null
  Time (mean ± σ):     10.004 s ±  0.144 s    [User: 9.425 s, System: 0.498 s]
  Range (min … max):    9.813 s … 10.183 s    10 runs

Benchmark 4: git log --oneline --shortstat --diff-algorithm=patience -10000 v6.0 >/dev/null
  Time (mean ± σ):     11.150 s ±  0.178 s    [User: 10.677 s, System: 0.387 s]
  Range (min … max):   11.009 s … 11.562 s    10 runs

Summary
  'git log --oneline --shortstat --diff-algorithm=myers -10000 v6.0 >/dev/null' ran
    1.04 ± 0.02 times faster than 'git log --oneline --shortstat --diff-algorithm=histogram -10000 v6.0 >/dev/null'
    1.11 ± 0.02 times faster than 'git log --oneline --shortstat --diff-algorithm=minimal -10000 v6.0 >/dev/null'
    1.11 ± 0.02 times faster than 'git log --oneline --shortstat --diff-algorithm=patience -10000 v6.0 >/dev/null'

Signed-off-by: Elijah Newren <newren@gmail.com>