xdiff/histogram: uniqueness of target file matters too
Our histogram diff was ported from JGit, where it was described as[1]:

    An extended form of Bram Cohen's patience diff algorithm.

    This implementation was derived by using the 4 rules that are
    outlined in Bram Cohen's blog, and then was further extended to
    support low-occurrence common elements.

    ...

    this algorithm behaves exactly like Bram Cohen's patience diff
    whenever there is a unique common element available between the two
    sequences.

Although the histogram diff was intended to be an extension of patience
diff, it failed to achieve that in practice.  On important difference
was that while patience diff computed both the lines unique to the
pre-image file and the lines unique to the post-image file, histogram
diff only computed occurrence frequencies for lines from the pre-image
file.  This meant that lines that aren't very unique on one side could
be preferred over actually unique lines in histogram diff.  This could
cause a surprise to users when they see that forward and backward diffs
are sometimes highly asymmetric.

Modify the histogram algorithm to compute frequencies of lines in the
post-image file as well.

Signed-off-by: Elijah Newren <newren@gmail.com>