xdiff/histogram: avoid recomputing histindex unnecessarily
When histogram diff recurses, it recomputes the histogram for the
remaining lines of interest.  As the remaining regions get smaller, the
frequency counts of remaining lines decreases.  Eventually, a blank line
or a line with just an opening or closing brace become "unique".  The
purpose of the histogram diff was to treat lower frequency lines as more
important to match on, so if we know we still have unique lines (which
have a repeat count of 2 -- appearing once on side A and once on side
B), then we should use those remaining unique lines before starting to
match on other lines.  Doing so is more consistent with the intent of
the algorithm.  (Note that we do eventually want high frequency lines
like spaces or braces to appear unique, because if we end up with a tiny
region, it makes sense to match on any remaining lines that are similar
instead of just emitting all removals and adds for the entire region.)

Avoiding recomputing frequency counts also provides a small performance
improvement of about 1-2% (I expected more, but perhaps the extra checks
in try_lcs() eat up the gains I thought I would get.)

Signed-off-by: Elijah Newren <newren@gmail.com>