replay: offer an option to linearize the commits
One of the stated goals of `git replay` is to allow implementing
the `git rebase` functionality on the server side.

Since the default mode of `git rebase` is to drop merge commits instead
of replaying them (linearizing the regular single-parent commits),
let's offer a mode to let `git replay` do the same.

Signed-off-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
[en: Adjust the patch for modern git, and change the testcase to not
 depend on other flags that have not been upstreamed]
Signed-off-by: Elijah Newren <newren@gmail.com>