checkout: add safety during in-progress operations
Back in commit c45f0f525de ("switch: reject if some operation is in
progress", 2019-03-29), considerable safety was added to `git switch` in
the form of throwing an error when attempting to switch to another
branch if you are in the middle of some type of operation (rebase,
merge, revert, cherry-pick).  Allowing the operation to proceed could
result in a very confusing mess for the user, so disallowing it saves
them from potential huge problems.

I noted when this was added to `git switch` that it should be added to
`git checkout` too[1], but at the time, Duy wanted to make it clear that
he was only adding a new subcommand, not modifying an existing one.  So,
fixing `git checkout` was left for future work, and has been on my TODO
list ever since.

As with the original commit, this commit does not implement an
`--ignore-in-progress` switch.  There is at least one case where doing
so might be safe: when in the middle of a merge and the target branch
equals HEAD[2].  However, there are some complications facing such an
option[3], as well as work needed to determine which set of cases that
flag should be able to operate under (and which still need to provide an
error message, though perhaps a more detailed one[4]), so we continue to
leave such a hypothetical option as future work.  For now, just add to
`git checkout` the middle-of-an-operation safety that `git switch` has.

[1] https://lore.kernel.org/git/CABPp-BHX1gRhTdurAwrPg60Hk-OuhbrEN=4zatx4OOUo-DkQvw@mail.gmail.com/
[2] https://lore.kernel.org/git/xmqqpm7k6ojz.fsf@gitster.g/
[3] https://lore.kernel.org/git/CABPp-BEksf4SuD57YsUO3YKhU12CAwFTy6pA1tETFrHB1DAz9w@mail.gmail.com/
[4] https://lore.kernel.org/git/CABPp-BGmPKyNcDa-wUh-oisTvvux+X=6BvGxSNQC2O7uodpFrA@mail.gmail.com/

Signed-off-by: Elijah Newren <newren@gmail.com>