checkout/switch: do not recommend possibly harmful/confusing commands
When --quit was added to the merge, am, rebase, cherry-pick, and revert
commands, it was explained with the use case that the user completely
forgot they were in the middle of an operation (this was first focused
around rebase), but continued building on top of that state anyway and
they no longer want to continue the operation, nor abort it and lose
the extra changes and commits they made while they had forgotten about
being in the middle of another operation.  (See commit 9512177b682
("rebase: add --quit to cleanup rebase, leave everything else
untouched", 2016-11-12), it was later added to other subcommands after
rebase.)

I believe that the use of --quit should be rare, and that more often,
users will remember that they are in the middle of some operation
(especially since modified command line prompts, `git status`,
checkout/switch, and perhaps other commands will all warn users about
being in the middle of an operation).  As such, we should guide users
towards using --abort instead.

This is especially important in the case of merge (and somewhat
important in the case of revert), where doing a --quit leaves around
changes that are much less likely to make sense as part of a simple
commit and which are very unlikely to be well described by a subsequent
commit message.

To address all of this, add a warning where --quit is documented that
users probably want to consider using --abort instead.

And, even more importantly, remove the confusing (and potentially even
dangerous) suggestion in checkout and switch for users to run
   git {merge,am,rebase,cherry-pick,revert} --quit
when in the middle of an operation, and suggest --abort instead.

Signed-off-by: Elijah Newren <newren@gmail.com>