sequencer: rename FROM_REBASE_PICK to FROM_CHERRY_PICK_DURING_REBASE
sequencer_determine_whence() tries to determine if the commit being
operated on came from a sequencer operation, and, if so, what kind of
operation it was.  It handles three cases:
  * FROM_CHERRY_PICK_SINGLE
  * FROM_CHERRY_PICK_MULTI
  * FROM_REBASE_PICK
The first two of these are exactly the cases you'd expect given their
names.  The third is extremely misleading; it is only ever set if both
CHERRY_PICK_HEAD and REBASE_HEAD are set, i.e. if we're in the middle of
a cherry-pick operation that was started while in the middle of a
rebase.  For simpler, normal rebase cases where CHERRY_PICK_HEAD is not
set, whence always end up being FROM_COMMIT (see determine_whence() in
builtin/commit.c).  So, the name is FROM_REBASE_PICK is grossly
misleading.

While it is also problematic that whence cannot track when we're in the
middle of other rebase operations, that will be handled in a future
commit.  For now, just rename FROM_REBASE_PICK to
FROM_CHERRY_PICK_DURING_REBASE to reflect its actual meaning and
purpose.

Signed-off-by: Elijah Newren <newren@gmail.com>