sequencer: simplify/clarify is_from_cherry_pick() and is_from_rebase()
As noted in the previous commit, the only types of operations
sequencer_determine_whence() will determine we are in the middle of are
all cherry-pick-related operations.  Even though the code split the
FROM_CHERRY_PICK_DURING_REBASE from the other two, the fact that all
three were really primarily cherry-pick-related operations is evidenced
by the fact that the code had several
   is_from_cherry_pick(w) || w == FROM_CHERRY_PICK_DURING_REBASE
code snippets.  Since a whence value of FROM_CHERRY_PICK_DURING_REBASE
is treated as a cherry-pick operation, adjust is_from_cherry_pick() and
is_from_rebase() accordingly and simplify the code.

The modification to is_from_rebase() technically results in two small
behavioral changes:
  1) If doing a partial commit we say that we cannot do it "during a
     cherry-pick" rather than "during a rebase" for
     FROM_CHERRY_PICK_DURING_REBASE.  Since both a cherry-pick and a
     rebase are in progress, this feels like a valid change.
  2) The fallback reflog message is "commit (cherry-pick)" instead of
     "commit (rebase)" when whence is FROM_CHERRY_PICK_DURING_REBASE.
     Again, since the cherry-pick is intended to be completed before
     the rebase, this seems like a bugfix.
Granted, we don't want is_from_rebase() to return false as it now does,
but we are only making the code be clearer that it doesn't correctly
handle middle-of-rebase operations.  Subsequence commits will fix this.

Signed-off-by: Elijah Newren <newren@gmail.com>