rev-list-options.adoc: correct one --ancestry-path example
Commit 257418c59040 (revision: allow --ancestry-path to take an
argument, 2022-08-19) added the possibility to pass a commit as an
argument to --ancestry-path (i.e. --ancestry-path=$COMMIT), with the
meaning that it would limit the output to commits that satisfy one of
the following
    * are an ancestor of $COMMIT
    * have $COMMMIT as an ancestor
    * are $COMMIT
and provided some documentation.  However, in the
    --ancestry-path=H D..M
example in the documentation, C is also within the range D..M and is an
ancestor of H, so it should also be shown.  Git does show it, but I
missed it when writing up the documentation.

Reported-by: Han Jiang <jhcarl0814@gmail.com>
Signed-off-by: Elijah Newren <newren@gmail.com>