replay: Smarter handling of new base
When I was replaying a commit in the middle of a history of the form
   A - B - C - F - G - I - J  <-- main
     \                /
       D - E - H -----

and I was doing a replay edit of E, afterwards I was noticing that I'd
get

   A     -------------- I* - J*
     \  /              /
       D - E* - H* ----

which was messed up.  If the either the first or second parents of a
merge are part of the UNINTERESTING set, then we've walked those and
they should map back to onto.  But if either of those parents are not
part of the UNINTERESTING set, then we didn't even walk them and we
should instead preserve those parents as-is.