replay: preserve merge parents unless found in replayed_commits
When I was replaying a commit in the middle of a history of the form
   A - B - C - F - G - I - J  <-- main
     \                /
       D - E - H -----

and I was doing a replay edit of E, afterwards I was noticing that I'd
get

   A     -------------- I* - J*
     \  /              /
       D - E* - H* ----

which was messed up.  I really want to preserve all parents of a merge
unless they are found in the replayed_commits remapping.

Merges can still be rebased, both because one of its parents may have
been replayed, and also because as a preliminary step, I insert into the
replay_commits remapping a map from each bottom commit to the 'onto'
commit.